参照機能(-rオプション)をDbOperatorクラスへ切り出した
・各オプションの処理をできる限りクラスとして切り出したいため、テストとして参照機能を切り出した。
